feat: blueprint service protocol
